Featrue: complete the fuction of factory
